+++ include/matrix.h
@@ -14,7 +14,7 @@ typedef struct
     size_t cols;
     size_t row_stride;
     size_t col_stride;
-    float *elements;
+    double *elements;
 } matrix_t;
 
 matrix_t matrix_alloc(size_t rows, size_t cols);
@@ -24,5 +24,5 @@ void matrix_copy(matrix_t dst, matrix_t src);
 
 void matrix_randomize_xavier(matrix_t matrix);
 void matrix_subtract(matrix_t minuend, matrix_t subtrahend);
-void matrix_scale(matrix_t matrix, float scalar);
+void matrix_scale(matrix_t matrix, double scalar);
 void matrix_element_multiply(matrix_t dst, matrix_t src);
