+++ include/vector.h
@@ -11,12 +11,12 @@
 typedef struct
 {
     size_t count;
-    float *elements;
+    double *elements;
 
 } vector_t;
 
 vector_t vector_alloc(size_t count);
 void vector_free(vector_t *vector);
-float vector_dot(vector_t v1, vector_t v2);
+double vector_dot(vector_t v1, vector_t v2);
 void vector_add(vector_t dst, vector_t lhs, vector_t rhs);
 void vector_copy(vector_t dst, vector_t src);
